[Add, Refactor] #8 - (BaseView) 추가 및 리팩토링
