+++ 35-seminar/Presentation/Practice/Week3/Week3PhotoViewController.swift
@@ -7,40 +7,38 @@
 
 import UIKit
 
-class Week3PhotoViewController: UIViewController {
-    
-    // MARK: - Properties
-    let appData = App.sampleApps
-    
-    let collectionView = UICollectionView()
-    
-    //    private var dataSource = UICollectionViewDiffableDataSource<PhotoSection, Photo>
-    
-    // MARK: - Methods
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setTableView()
-    }
-    
-    private func setTableView() {
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        collectionView.register(Week3PhotoCollectionViewCell.self, forCellWithReuseIdentifier: Week3PhotoCollectionViewCell.identifier)
-    }
-}
-
-extension Week3PhotoViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        <#code#>
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        <#code#>
-    }
-    
-    
-}
-
-extension Week3PhotoViewController: UICollectionViewDelegate {
-    
-}
+//class Week3PhotoViewController: UIViewController {
+//    
+//    // MARK: - Properties
+//    let appData = App.sampleApps
+//    
+//    let collectionView = UICollectionView()
+//    
+//    //    private var dataSource = UICollectionViewDiffableDataSource<PhotoSection, Photo>
+//    
+//    // MARK: - Methods
+//    override func viewDidLoad() {
+//        super.viewDidLoad()
+//        setTableView()
+//    }
+//    
+//    private func setTableView() {
+//        collectionView.dataSource = self
+//        collectionView.delegate = self
+//        collectionView.register(Week3PhotoCollectionViewCell.self, forCellWithReuseIdentifier: Week3PhotoCollectionViewCell.identifier)
+//    }
+//}
+//
+//extension Week3PhotoViewController: UICollectionViewDataSource {
+//    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+//        <#code#>
+//    }
+//    
+//    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+//        <#code#>
+//    }
+//}
+//
+//extension Week3PhotoViewController: UICollectionViewDelegate {
+//    
+//}
